+++ R/tableby.control.R
@@ -29,7 +29,7 @@
 #' @param selectall.test name of test for date variables: notest
 #' @param stats.labels A named list of labels for all the statistics function names, where the function name is the named element in the list
 #'   and the value that goes with it is a string containing the formal name that will be printed in all printed renderings of the output,
-#'   e.g., \code{list(countpct="Count (Pct)")}.
+#'   e.g., \code{list(countpct="Count (Pct)")}. Any unnamed elements will be ignored. Passing \code{NULL} will disable labels.
 #' @param digits Number of decimal places for numeric values.
 #' @param digits.count Number of decimal places for count values.
 #' @param digits.pct Number of decimal places for percents.
@@ -82,12 +82,7 @@ tableby.control <- function(
   numeric.stats=c("Nmiss","meansd","range"), cat.stats=c("Nmiss","countpct"),
   ordered.stats=c("Nmiss", "countpct"), surv.stats=c("Nmiss", "Nevents","medSurv"), date.stats=c("Nmiss", "median","range"),
   selectall.stats=c("Nmiss", "countpct"),
-  stats.labels=list(Nmiss="N-Miss", Nmiss2="N-Miss", meansd="Mean (SD)", medianrange="Median (Range)",
-                    median="Median", medianq1q3="Median (Q1, Q3)", q1q3="Q1, Q3", iqr = "IQR",
-                    mean = "Mean", sd = "SD", var = "Var", max = "Max", min = "Min", meanCI = "Mean (CI)", sum = "Sum",
-                    gmean = "Geom Mean", gsd = "Geom SD", gmeansd = "Geom Mean (Geom SD)", gmeanCI = "Geom Mean (CI)",
-                    range="Range", Npct="N (Pct)", Nevents="Events", medSurv="Median Survival",
-                    medTime = "Median Follow-Up", medianmad="Median (MAD)", overall = "Overall", total = "Total", difference = "Difference"),
+  stats.labels = list(),
   digits = 3L, digits.count = 0L, digits.pct = 1L, digits.p = 3L, format.p = TRUE, digits.n = 0L, conf.level = 0.95,
   chisq.correct=FALSE, simulate.p.value=FALSE, B=2000, times = 1:5, ...) {
 
@@ -124,6 +119,8 @@ tableby.control <- function(
     digits.n <- 0L
   }
 
+  stats.labels <- if(is.null(stats.labels)) NULL else add_tbc_stats_labels(stats.labels)
+
   list(test=test, total=total, test.pname=test.pname,
        numeric.simplify=numeric.simplify, cat.simplify=cat.simplify, ordered.simplify=ordered.simplify, date.simplify=date.simplify,
        numeric.test=numeric.test, cat.test=cat.test, ordered.test=ordered.test, surv.test=surv.test, date.test=date.test, selectall.test=selectall.test,
@@ -132,4 +129,19 @@ tableby.control <- function(
        stats.labels=stats.labels,
        digits=digits, digits.p=digits.p, digits.count = digits.count, digits.pct = digits.pct, format.p = format.p, digits.n = digits.n,
        conf.level=conf.level, chisq.correct=chisq.correct, simulate.p.value=simulate.p.value, B=B, times=times)
+}
+
+add_tbc_stats_labels <- function(x) {
+  start <- list(
+    Nmiss="N-Miss", Nmiss2="N-Miss", meansd="Mean (SD)", medianrange="Median (Range)",
+    median="Median", medianq1q3="Median (Q1, Q3)", q1q3="Q1, Q3", iqr = "IQR",
+    mean = "Mean", sd = "SD", var = "Var", max = "Max", min = "Min", meanCI = "Mean (CI)", sum = "Sum",
+    gmean = "Geom Mean", gsd = "Geom SD", gmeansd = "Geom Mean (Geom SD)", gmeanCI = "Geom Mean (CI)",
+    range="Range", Npct="N (Pct)", Nevents="Events", medSurv="Median Survival",
+    medTime = "Median Follow-Up", medianmad="Median (MAD)",
+    overall = "Overall", total = "Total", difference = "Difference"
+  )
+  nms <- setdiff(names(x), "")
+  start[nms] <- x[nms]
+  start
 }